requirements.txt
pypdf>=6.0.0
tabula-py>=2.10.0
pandas>=2.3.2